move load_json to io.py
